+++ FrontEnd/src/Pages/LogIn.css
@@ -4,9 +4,6 @@
   padding: 0;
 }
 
-.header{
-  color:black;
-}
 
 .general{
   display: flex;
@@ -14,12 +11,40 @@
   justify-content: space-between;
   text-align: left;
   margin: 3% 15%;
+
 }
 
-
-.title{
+.general-title{
   text-align: left;
 }
+
+.both-sides{
+  display: flex;
+  flex-direction: row;
+  margin: 7% 0px;
+}
+
+.left-side{
+  border-right: solid;
+  border-color: black;
+  border-width: 1px;
+  width: 50%;
+  padding: 0px 5vh;
+  
+}
+
+.right-side{
+  text-align: left;
+  padding-left: 10%;
+  padding-right: 10%;
+  width: 50%;
+}
+
+.right-side-subtitle{
+  font-size: 20px;
+  font-weight: 700;
+}
+
 
 
 .email {
@@ -31,7 +56,7 @@
   justify-content: center;
 }
 
-.email-subscribe-li {
+.email-subscribe-si {
   padding: 15px 10px;
   border: solid;
   font-size: 15px;
@@ -40,63 +65,50 @@
   border-width: 1px;
   border-color: rgb(204, 204, 204);
   margin: 8px 5px;
+  width: 90%;
 }
 
-.eemail-subscribe-li {
-  padding: 15px 10px;
-  border: solid;
-  font-size: 15px;
-  font-weight: bold;
-  outline: none;
-  border-width: 1px;
-  border-color: rgb(204, 204, 204);
-  margin: 8px 5px;
-  width: 470px;
-}
 
-.email-subscribe-li::placeholder {
+
+.email-subscribe-si::placeholder {
   color: black;
   font-weight: bold;
 }
 
-.eemail-subscribe-li::placeholder {
-  color: black;
-  font-weight: bold;
-}
-
-
-.email-subscribe-li:focus{
+.email-subscribe-si:focus{
   border-color: black;
 }
 
-.email-subscribe-li:hover{
+.email-subscribe-si:hover{
   border-color: black;
 }
 
-.eemail-subscribe-li:focus{
-  border-color: black;
-}
 
-.eemail-subscribe-li:hover{
-  border-color: black;
-}
 
-.pass-char{
-  font-size: 11px;
-}
-
-.want-suscribe{
-  display: flex;
-  flex-direction: row;
-}
-
-.btn-unirse-li {
+.btn-unirse-si {
   background-color: rgb(230, 230, 230);
   border: none;
   padding: 15px 10%;
   color: rgb(153, 153, 153);
   font-weight: bolder;
   margin: 5px;
+  width: 90%;
+}
+
+.btn-create-si {
+  background-color: rgb(0, 0, 0);
+  border: none;
+  padding: 15px 10%;
+  color: rgb(255, 255, 255);
+  font-weight: bolder;
+  margin: 5px;
+  width: 90%;
+
+  transition: all 0.3s ease;
+}
+
+.btn-create-si:hover{
+  opacity: .6;
 }
 
 .pass-inputs{
@@ -108,46 +120,82 @@
 .pass-input{
   position: relative;
 }
-.pass-icon-li{
+.pass-icon{
   position: absolute;
   top: 35%;
-  right: 10%;
+  right: 12%;
   cursor: pointer;
 }
 
-.or-signin{
+.or-reset-pass{
   color: rgb(0, 69, 255);
-  font-size: smaller;
+  font-size: 12px;
+  margin-top: 20px;
+  font-weight: bold;
 }
 
 
 
 
 @media (max-width: 800px) {
-  
-
-  .pass-icon-li{
-    position: absolute;
-    top: 35%;
-    right: 59%;
-    cursor: pointer;
+  .both-sides {
+    flex-direction: column;
   }
-  .contenedor-footer {
-      width: 100%;
-      margin: 5px 0;
-      flex-basis: calc(25% - 20px); /* Utilizar el 100% del espacio en flex basis */
+ 
+  .left-side{
+    border-right: none;
+    border-bottom: solid;
+    border-color: black;
+    border-width: 1px;
+    width: 98%;
+    margin: 0px 15px;
+    text-align: center;
+  }
+  .pass-icon{
+    right: 10%;
+  }
+
+  .both-sides {
+    flex-direction: column;
+  }
+  .right-side{
+    text-align: left;
+    padding-left: 10%;
+    padding-top: 20px;
+  }
+
+  .pass-inputs {
+    flex-direction: column;
+  }
+  .pass-icon{
+    right: 15%;
+  }
+  
+ .email-subscribe-si{
+    width: 90%;
+  }
+
+  .btn-unirse-si{
+    width: 90%;
   }
 }
 
-@media (max-width: 800px) {
+
+
+@media (max-width: 600px) {
+  .both-sides {
+    flex-direction: column;
+  }
+
+  .pass-inputs {
+    flex-direction: column;
+  }
+  .pass-icon{
+    right: 15%;
+  }
   
-  .eemail-subscribe-li {
-    
-    width: 230px;
+  .eemail-subscribe-si{
+    width: 90%;
   }
-  .contenedor-footer {
-      width: 100%;
-      margin: 5px 0;
-      flex-basis: calc(25% - 20px); /* Utilizar el 100% del espacio en flex basis */
-  }
+
 }
